fix(users bathingspots put): fixes PUT by using the bathingspot-helpers
