+++ .snyk
@@ -1,0 +1,9 @@
+# Snyk (https://snyk.io) policy file, patches or ignores known vulnerabilities.
+version: v1.13.3
+# ignores vulnerabilities until expiry date; change duration by modifying expiry date
+ignore:
+  'npm:mem:20180117':
+    - typeorm > cli-highlight > yargs > os-locale > mem:
+        reason: No never version available
+        expires: '2019-05-01T06:24:10.045Z'
+patch: {}
